+++ src/components/victory-area/victory-area.jsx
@@ -1,4 +1,3 @@
-import pick from "lodash/pick";
 import last from "lodash/last";
 import assign from "lodash/assign";
